requirements.txt
fastapi[standard]
uvicorn
mysql-connector-python
python-dotenv
python-multipart
Jinja2
python-jose
passlib
sqlalchemy
pydantic
